CR3: Fix bugs and implement a main runner to execute all steps of the dataset creation pipeline
