+++ requirements.txt
@@ -16,4 +16,5 @@ transformers
 gitpython
 bitsandbytes
 accelerate>=0.26.0
+ollama
